Fix invalid conversion.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/maps/trunk@21951 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ python/maps/darcap/manor.py
@@ -91,7 +91,7 @@ def kaptel_death():
         return
 
     who.Map.Print("%s roars and seems to regenerate!"%(Crossfire.WhoAmI().Name))
-    who.HP = who.MaxHP / 2
+    who.HP = int(who.MaxHP / 2)
     Crossfire.SetReturnValue(1)
 
 def challenge_correct_reply(count, msg):
